tradingbot_final_new/requirements.txt
fastapi
uvicorn
python-binance>=1.0.19
requests
pandas
numpy
python-dotenv
